Separate the configuration file structs from the parsing functions.
