+++ crates/cli/src/model.rs
@@ -1,4 +1,3 @@
 pub mod cli_configuration;
 pub mod config_file;
 pub mod datadog_api;
-mod serialization;
